requirements.txt
pytest
imgaug
pandas
pillow